feat(reports): sweep_report generator with provenance banner + archive
Moves the workflow-sweep HTML report into scripts/ per AGENTS.md's "HTML
reports with provenance banners" convention. Reads a sweep's hive-partitioned
parquet and writes two self-contained files (plots base64-embedded): a latest
copy and a timestamp+commit archival copy under reports/figures/<experiment>/.

Provenance banner matches scripts/compare_pdmp_vs_baseline.py::collect_provenance
— ISO-8601 timestamp, full+short SHA linked to the GitHub commit, branch,
DIRTY TREE badge, last-commit msg/author/date, generator path, host/OS/Python.
Embeds how-to-run docs (seeds/generations/variants knobs + CLI) and a summary
of the workflow framework. Generalizes over variants × seeds × generations.

  .venv/bin/python scripts/sweep_report.py <sweep_dir> --open

Replaces the earlier reports/sweep_report.py (wrong location, thinner banner).
